Refactor math expression parser code to support string values

+++ src/main/java/mchorse/mclib/client/gui/framework/elements/input/GuiTrackpadElement.java
@@ -401,7 +401,7 @@ public class GuiTrackpadElement extends GuiBaseTextElement
                 {
                     MathBuilder builder = new MathBuilder();
 
-                    this.setValueAndNotify(builder.parse(this.field.getText()).get());
+                    this.setValueAndNotify(builder.parse(this.field.getText()).get().doubleValue());
                 }
                 catch (Exception e)
                 {}
